feat: call rules during fx quote
