+++ src/model/quotes.js
@@ -52,8 +52,7 @@ const { httpRequest } = require('../lib/http')
 const { getStackOrInspect, generateRequestHeadersForJWS, generateRequestHeaders, calculateRequestHash, fetchParticipantInfo, getParticipantEndpoint } = require('../lib/util')
 const { RESOURCES } = require('../constants')
 const LOCAL_ENUM = require('../lib/enum')
-const rules = require('../../config/rules.json')
-const RulesEngine = require('./rules.js')
+const { executeRules, handleRuleEvents } = require('./executeRules')
 
 delete axios.defaults.headers.common.Accept
 delete axios.defaults.headers.common['Content-Type']
@@ -75,79 +74,11 @@ class QuotesModel {
     })
   }
 
-  async executeRules (headers, quoteRequest, payer, payee) {
-    if (rules.length === 0) {
-      return []
-    }
-
-    const facts = {
-      payer,
-      payee,
-      payload: quoteRequest,
-      headers
-    }
-
-    const { events } = await RulesEngine.run(rules, facts)
-    this.log.debug('Rules engine returned events:', { events })
-
-    return events
+  executeRules () {
+    return executeRules.apply(this, arguments)
   }
 
-  async handleRuleEvents (events, headers, payload, originalPayload) {
-    const quoteRequest = originalPayload || payload
-    // todo: pass only originalPayload (added this logic only for passing tests)
-
-    // At the time of writing, all events cause the "normal" flow of execution to be interrupted.
-    // So we'll return false when there have been no events whatsoever.
-    if (events.length === 0) {
-      return { terminate: false, quoteRequest, headers }
-    }
-
-    const { INVALID_QUOTE_REQUEST, INTERCEPT_QUOTE } = RulesEngine.events
-
-    const unhandledEvents = events.filter(ev => !(ev.type in RulesEngine.events))
-
-    if (unhandledEvents.length > 0) {
-      // The rules configuration contains events not handled in the code
-      // TODO: validate supplied rules at startup and fail if any invalid rules are discovered.
-      throw new Error('Unhandled event returned by rules engine')
-    }
-
-    const invalidQuoteRequestEvents = events.filter(ev => ev.type === INVALID_QUOTE_REQUEST)
-    if (invalidQuoteRequestEvents.length > 0) {
-      // Use the first event, ignore the others for now. This is ergonomically worse for someone
-      // developing against this service, as they can't see all reasons their quote was invalid at
-      // once. But is a valid solution in the short-term.
-      const { FSPIOPError: code, message } = invalidQuoteRequestEvents[0].params
-      // Will throw an internal server error if property doesn't exist
-      throw ErrorHandler.CreateFSPIOPError(ErrorHandler.Enums.FSPIOPErrorCodes[code],
-        message, null, headers['fspiop-source'])
-    }
-
-    const interceptQuoteEvents = events.filter(ev => ev.type === INTERCEPT_QUOTE)
-    if (interceptQuoteEvents.length > 1) {
-      // TODO: handle priority. Can we stream events?
-      throw new Error('Multiple intercept quote events received')
-    }
-    if (interceptQuoteEvents.length > 0) {
-      // send the quote request to the recipient in the event
-      const result = {
-        terminate: false,
-        quoteRequest,
-        headers: {
-          ...headers,
-          'fspiop-destination': interceptQuoteEvents[0].params.rerouteToFsp
-        }
-      }
-      // if additionalHeaders are present then add the additional non-standard headers (e.g. used by forex)
-      // Note these headers are not part of the mojaloop specification
-      if (interceptQuoteEvents[0].params.additionalHeaders) {
-        result.headers = { ...result.headers, ...interceptQuoteEvents[0].params.additionalHeaders }
-        result.additionalHeaders = interceptQuoteEvents[0].params.additionalHeaders
-      }
-      return result
-    }
-  }
+  handleRuleEvents = handleRuleEvents
 
   /**
    * Validates the quote request object
@@ -281,9 +212,8 @@ class QuotesModel {
 
       // Run the rules engine. If the user does not want to run the rules engine, they need only to
       // supply a rules file containing an empty array.
-      const events = await this.executeRules(headers, quoteRequest, payer, payee)
+      handledRuleEvents = await this.executeRules(headers, quoteRequest, originalPayload, payer, payee, 'quoteRequest')
 
-      handledRuleEvents = await this.handleRuleEvents(events, headers, quoteRequest, originalPayload)
       if (handledRuleEvents.terminate) {
         return
       }
